+++ backend/middleware.py
@@ -14,12 +14,26 @@ def send_error(environ, start_response):
     return response(environ, start_response)
 
 
+# to help render the apidocs endpoint with flasgger
+SWAGGER_UI_ROUTES = [
+    "/flasgger_static/swagger-ui.css",
+    "/flasgger_static/swagger-ui-bundle.js",
+    "/flasgger_static/swagger-ui-standalone-preset.js",
+    "/flasgger_static/lib/jquery.min.js",
+    "/flasgger_static/favicon-32x32.png",
+]
+
+
 class middleware:
-    def __init__(self, app, exempt_paths=[]):
+    def __init__(self, app, exempt_paths=None):
         self.app = app
+        self.exempt_paths = exempt_paths if exempt_paths else []
 
     def __call__(self, environ, start_response):
         request = Request(environ)
+        if request.path in SWAGGER_UI_ROUTES or request.path in self.exempt_paths:
+            print("in exempt paths")
+            return self.app(environ, start_response)
         if (
             "Authorization" in request.headers
             and "Bearer " in request.headers["Authorization"]
